Adding s3 client

+++ libs/libcommon/src/libcommon/viewer_utils/asset.py
@@ -2,7 +2,6 @@
 # Copyright 2022 The HuggingFace Authors.
 
 import contextlib
-import io
 import json
 import os
 from hashlib import sha1
@@ -11,10 +10,10 @@ from pathlib import Path
 from typing import Generator, List, Optional, Tuple, TypedDict
 from uuid import uuid4
 
-import boto3
 from PIL import Image  # type: ignore
 from pydub import AudioSegment  # type:ignore
 
+from libcommon.s3_client import S3Client
 from libcommon.storage import StrPath
 
 DATASET_SEPARATOR = "--"
@@ -83,11 +82,9 @@ def create_image_file(
     overwrite: bool = True,
     # TODO: Once assets and cached-assets are migrated to S3, this parameter is no more needed
     use_s3_storage: bool = False,
+    s3_client: Optional[S3Client] = None,
     # TODO: Once assets and cached-assets are migrated to S3, the following parameters dont need to be optional
     s3_bucket: Optional[str] = None,
-    s3_access_key_id: Optional[str] = None,
-    s3_secret_access_key: Optional[str] = None,
-    s3_region: Optional[str] = None,
     s3_folder_name: Optional[str] = None,
 ) -> ImageSource:
     dir_path, url_dir_path = get_asset_dir_name(
@@ -98,33 +95,32 @@ def create_image_file(
         column=column,
         assets_directory=assets_directory,
     )
-    if use_s3_storage:
-        key = f"{s3_folder_name}/{url_dir_path}/{filename}"
-        s3_client = boto3.client(
-            S3_RESOURCE,
-            region_name=s3_region,
-            aws_access_key_id=s3_access_key_id,
-            aws_secret_access_key=s3_secret_access_key,
-        )
-        exists = True
-        if not overwrite:
-            try:
-                s3_client.head_object(
-                    Bucket=s3_bucket,
-                    Key=key,
-                )
-            except Exception:
-                exists = False
-        if overwrite or not exists:
-            image_byte_arr = io.BytesIO()
-            image.save(image_byte_arr, format="JPEG")
-            image_byte_arr.seek(0)
-            s3_client.upload_fileobj(image_byte_arr, s3_bucket, key)
-    else:
-        makedirs(dir_path, ASSET_DIR_MODE, exist_ok=True)
+
+    if not use_s3_storage:
         file_path = dir_path / filename
-        if overwrite or not file_path.exists():
-            image.save(file_path)
+        makedirs(dir_path, ASSET_DIR_MODE, exist_ok=True)
+    else:
+        payload = (
+            str(uuid4()),
+            dataset,
+            config,
+            split,
+            row_idx,
+            column,
+        )
+        prefix = sha1(json.dumps(payload, sort_keys=True).encode(), usedforsecurity=False).hexdigest()[:8]
+        file_path = Path(assets_directory).resolve() / f"{prefix}-{filename}"
+
+    if overwrite or not file_path.exists():
+        image.save(file_path)
+
+    if use_s3_storage and s3_client is not None and s3_bucket is not None:
+        key = f"{s3_folder_name}/{url_dir_path}/{filename}"
+        create_object = overwrite or not s3_client.exists_in_bucket(s3_bucket, key)
+        if create_object:
+            s3_client.upload_to_bucket(str(file_path), s3_bucket, key)
+            os.remove(file_path)
+
     src = f"{assets_base_url}/{url_dir_path}/{filename}"
     return {
         "src": src,
@@ -151,11 +147,9 @@ def create_audio_file(
     overwrite: bool = True,
     # TODO: Once assets and cached-assets are migrated to S3, this parameter is no more needed
     use_s3_storage: bool = False,
+    s3_client: Optional[S3Client] = None,
     # TODO: Once assets and cached-assets are migrated to S3, the following parameters dont need to be optional
     s3_bucket: Optional[str] = None,
-    s3_access_key_id: Optional[str] = None,
-    s3_secret_access_key: Optional[str] = None,
-    s3_region: Optional[str] = None,
     s3_folder_name: Optional[str] = None,
 ) -> List[AudioSource]:
     dir_path, url_dir_path = get_asset_dir_name(
@@ -186,27 +180,11 @@ def create_audio_file(
         segment: AudioSegment = AudioSegment.from_file(audio_file_path)
         segment.export(file_path, format="mp3")
 
-    if use_s3_storage:
+    if use_s3_storage and s3_client is not None and s3_bucket is not None:
         mp3_key = f"{s3_folder_name}/{url_dir_path}/{filename}"
-
-        s3_client = boto3.client(
-            S3_RESOURCE,
-            region_name=s3_region,
-            aws_access_key_id=s3_access_key_id,
-            aws_secret_access_key=s3_secret_access_key,
-        )
-        mp3_exists = True
-        if not overwrite:
-            try:
-                s3_client.head_object(
-                    Bucket=s3_bucket,
-                    Key=mp3_key,
-                )
-            except Exception:
-                mp3_exists = False
-
-        if overwrite or not mp3_exists:
-            s3_client.upload_file(file_path, s3_bucket, mp3_key)
+        create_object = overwrite or not s3_client.exists_in_bucket(s3_bucket, mp3_key)
+        if create_object:
+            s3_client.upload_to_bucket(str(file_path), s3_bucket, mp3_key)
             os.remove(file_path)
 
     return [
